backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
python-multipart>=0.0.6
httpx>=0.25.0
pydantic>=2.4.0
python-dotenv>=1.0.0
python-jose>=3.3.0
twelvelabs>=1.1.0
httpx>=0.25.2
